mgmt: fix circular import
